app/requirements.txt
streamlit>=1.28.0
databricks-sdk>=0.20.0
